tests/wpt/tests/css/css-conditional/container-queries/scroll-state/scroll-state-snapped-change.html

<!DOCTYPE html>
<title>@container: scroll-state(snapped) changed after scroll</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  :root {
    scroll-snap-type: block mandatory;
  }
  body {
    margin: 0;
  }
  #filler {
    height: 10000px;
  }
  .snapped {
    container-type: scroll-state;
    scroll-snap-align: start;
    width: 100px;
    height: 100px;
    margin-bottom: 100px;
    background: teal;
  }

  @container scroll-state(snapped: block) {
    span { --snapped: yes }
  }
</style>
<div class="snapped">
  <span id="target1"></span>
</div>
<div class="snapped">
  <span id="target2"></span>
</div>
<div id="filler"></div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  promise_test(async t => {
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target1).getPropertyValue("--snapped"), "yes");
    assert_equals(getComputedStyle(target2).getPropertyValue("--snapped"), "");
    assert_equals(document.documentElement.scrollTop, 0);
  }, "Check that scroll-state(snapped: block) matches on #target1 before scroll");

  promise_test(async t => {
    document.documentElement.scrollTop = 300;
    await waitForAnimationFrames(2);
    assert_equals(getComputedStyle(target1).getPropertyValue("--snapped"), "");
    assert_equals(getComputedStyle(target2).getPropertyValue("--snapped"), "yes");
    assert_equals(document.documentElement.scrollTop, 200);
  }, "Check that scroll-state(snapped: block) matches on #target2 after scroll");
</script>
